+++ skyperious/conf.py
@@ -22,7 +22,7 @@ import sys
 
 """Program title, version number and version date."""
 Title = "Skyperious"
-Version = "4.0.dev31"
+Version = "4.0.dev33"
 VersionDate = "27.07.2020"
 
 if getattr(sys, "frozen", False):