+++ CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 ---------------
 - added support for syncing chat history from Skype online service;
 - added 400+ more emoticon images;
-- added date range option to command-line export;
+- added date range option to chat export;
 - restored shared image download functionality;
 - restored support functionality;
 - made database list sortable and filterable, added date and size columns;